#12 - added freeplane project & removed MindmapService from client side
