+++ org.flowerplatform.core/.classpath
@@ -3,11 +3,5 @@
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.7"/>
 	<classpathentry kind="con" path="org.eclipse.pde.core.requiredPlugins"/>
 	<classpathentry kind="src" path="src"/>
-	<classpathentry kind="lib" path="lib/freeplaneviewer.jar" sourcepath="C:/Users/PowerUser/git/freeplane"/>
-	<classpathentry kind="lib" path="lib/commons-lang-2.6.jar"/>
-	<classpathentry kind="lib" path="lib/gnu-regexp-1.1.4.jar"/>
-	<classpathentry kind="lib" path="lib/jortho.jar"/>
-	<classpathentry kind="lib" path="lib/SimplyHTML.jar"/>
-	<classpathentry kind="lib" path="lib/freeplaneeditor_modified.jar"/>
 	<classpathentry kind="output" path="bin"/>
 </classpath>
