+++ org.flowerplatform.core/META-INF/MANIFEST.MF
@@ -4,14 +4,11 @@ Bundle-Name: Core
 Bundle-SymbolicName: org.flowerplatform.core
 Bundle-Version: 1.0.0.qualifier
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
-Export-Package: org.flowerplatform.core.node.remote
-Bundle-ClassPath: lib/commons-lang-2.6.jar,
- lib/freeplaneviewer.jar,
- lib/freeplaneeditor_modified.jar,
- lib/gnu-regexp-1.1.4.jar,
- lib/jortho.jar,
- lib/SimplyHTML.jar,
- .
+Export-Package: org.flowerplatform.core,
+ org.flowerplatform.core.node,
+ org.flowerplatform.core.node.controller,
+ org.flowerplatform.core.node.remote
+Bundle-ClassPath: .
 Require-Bundle: org.flowerplatform.util,
  org.eclipse.osgi
 Bundle-Activator: org.flowerplatform.core.CorePlugin
