+++ org.flowerplatform.core/build.properties
@@ -1,10 +1,4 @@
 output.. = bin/
 bin.includes = META-INF/,\
-               lib/commons-lang-2.6.jar,\
-               lib/freeplaneeditor_modified.jar,\
-               lib/freeplaneviewer.jar,\
-               lib/gnu-regexp-1.1.4.jar,\
-               lib/jortho.jar,\
-               lib/SimplyHTML.jar,\
                .
 source.. = src/
