+++ org.flowerplatform.core/src/org/flowerplatform/core/CorePlugin.java
@@ -1,12 +1,5 @@
 package org.flowerplatform.core;
 
-import org.flowerplatform.core.mindmap.FreeplaneAddNodeController;
-import org.flowerplatform.core.mindmap.FreeplaneChildrenProvider;
-import org.flowerplatform.core.mindmap.FreeplanePropertiesProvider;
-import org.flowerplatform.core.mindmap.FreeplanePropertySetter;
-import org.flowerplatform.core.mindmap.FreeplaneRemoveNodeController;
-import org.flowerplatform.core.mindmap.FreeplaneUtils;
-import org.flowerplatform.core.node.NodeTypeDescriptor;
 import org.flowerplatform.core.node.NodeTypeDescriptorRegistry;
 import org.flowerplatform.core.node.remote.NodeService;
 import org.flowerplatform.util.plugin.AbstractFlowerJavaPlugin;
@@ -27,14 +20,7 @@ public class CorePlugin extends AbstractFlowerJavaPlugin {
 	public void start(BundleContext bundleContext) throws Exception {
 		super.start(bundleContext);
 		INSTANCE = this;
-		
-		NodeTypeDescriptor nodeTypeDescriptor = nodeTypeDescriptorRegistry.getOrCreateNodeTypeDescriptor("freeplaneNode");
-		nodeTypeDescriptor.addChildrenProvider(new FreeplaneChildrenProvider());
-		nodeTypeDescriptor.addPropertiesProvider(new FreeplanePropertiesProvider());
-		nodeTypeDescriptor.addAddNodeController(new FreeplaneAddNodeController());
-		nodeTypeDescriptor.addRemoveNodeController(new FreeplaneRemoveNodeController());
-		nodeTypeDescriptor.addPropertySetter(new FreeplanePropertySetter());
-		
+				
 		getServiceRegistry().registerService("nodeService", new NodeService(nodeTypeDescriptorRegistry));
 	}
 
@@ -56,17 +42,9 @@ public class CorePlugin extends AbstractFlowerJavaPlugin {
 	}
 	
 	/**
-	 * TODO CC: move to Freeplane plugin
 	 * @author Cristina Constantinescu
 	 */
-	private FreeplaneUtils freeplaneUtils = new FreeplaneUtils();
-
-	/**
-	 * @author Cristina Constantinescu
-	 */
-	public FreeplaneUtils getFreeplaneUtils() {
-		return freeplaneUtils;
+	public void registerMessageBundle() throws Exception {
+		// not used
 	}
-
-
 }
