+++ org.flowerplatform.flex_client.core/src/org/flowerplatform/flex_client/core/CorePlugin.as
@@ -17,11 +17,14 @@
  * license-end
  */
 package org.flowerplatform.flex_client.core {
-	import flash.net.registerClassAlias;
-	
-	import org.flowerplatform.flex_client.core.mindmap.MindMapPerspective;
-	import org.flowerplatform.flex_client.core.mindmap.MindMapService;
+	import org.flowerplatform.flex_client.core.mindmap.action.AddNodeAction;
+	import org.flowerplatform.flex_client.core.mindmap.action.RefreshAction;
+	import org.flowerplatform.flex_client.core.mindmap.action.ReloadAction;
+	import org.flowerplatform.flex_client.core.mindmap.action.RemoveNodeAction;
+	import org.flowerplatform.flex_client.core.mindmap.action.RenameAction;
+	import org.flowerplatform.flex_client.core.mindmap.action.SaveAction;
 	import org.flowerplatform.flex_client.core.mindmap.layout.MindMapEditorProvider;
+	import org.flowerplatform.flex_client.core.mindmap.layout.MindMapPerspective;
 	import org.flowerplatform.flex_client.core.mindmap.remote.Node;
 	import org.flowerplatform.flex_client.core.plugin.AbstractFlowerFlexPlugin;
 	import org.flowerplatform.flex_client.core.service.ServiceLocator;
@@ -39,8 +42,6 @@ package org.flowerplatform.flex_client.core {
 		protected static var INSTANCE:CorePlugin;
 		
 		public var serviceLocator:ServiceLocator = new ServiceLocator();
-		
-		public var mindMapService:MindMapService = new MindMapService();
 		
 		public var perspectives:Vector.<Perspective> = new Vector.<Perspective>();
 		
@@ -65,11 +66,18 @@ package org.flowerplatform.flex_client.core {
 			}
 			INSTANCE = this;
 				
-			serviceLocator.addService(MindMapService.ID);
+			serviceLocator.addService("nodeService");
+			serviceLocator.addService("freeplaneService");
 			
-			FlexUtilGlobals.getInstance().composedViewProvider.addViewProvider(new MindMapEditorProvider());
-			
+			FlexUtilGlobals.getInstance().composedViewProvider.addViewProvider(new MindMapEditorProvider());			
 			perspectives.push(new MindMapPerspective());
+			
+			mindmapEditorClassFactoryActionProvider.addActionClass(AddNodeAction);
+			mindmapEditorClassFactoryActionProvider.addActionClass(RemoveNodeAction);	
+			mindmapEditorClassFactoryActionProvider.addActionClass(RenameAction);	
+//			mindmapEditorClassFactoryActionProvider.addActionClass(ReloadAction);
+//			mindmapEditorClassFactoryActionProvider.addActionClass(RefreshAction);
+			mindmapEditorClassFactoryActionProvider.addActionClass(SaveAction);
 			
 //			linkHandlers = new Dictionary();			
 //
