+++ org.flowerplatform.flex_client.server/META-INF/MANIFEST.MF
@@ -6,7 +6,8 @@ Bundle-Version: 1.0.0.qualifier
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Require-Bundle: org.flowerplatform.util,
  org.eclipse.equinox.http.registry,
- org.flowerplatform.core
+ org.flowerplatform.core,
+ org.flowerplatform.freeplane
 Import-Package: javax.servlet,
  javax.servlet.http
 Bundle-ClassPath: lib/commons-codec-1.3.jar,
