+++ org.flowerplatform.flex_client.server/blazeds-services-config/remoting-config.xml
@@ -15,4 +15,10 @@
         </properties>
     </destination>
     
+     <destination id="freeplaneService">
+        <properties>
+            <source>org.flowerplatform.freeplane.remote.FreeplaneService</source>
+        </properties>
+    </destination>
+    
 </service>
