+++ org.flowerplatform.freeplane/META-INF/MANIFEST.MF
@@ -1,0 +1,18 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Freeplane
+Bundle-SymbolicName: org.flowerplatform.freeplane;singleton:=true
+Bundle-Version: 1.0.0.qualifier
+Require-Bundle: org.flowerplatform.core,
+ org.flowerplatform.util,
+ org.eclipse.osgi
+Bundle-ClassPath: lib/commons-lang-2.6.jar,
+ lib/freeplaneeditor_modified.jar,
+ lib/freeplaneviewer.jar,
+ lib/gnu-regexp-1.1.4.jar,
+ lib/jortho.jar,
+ lib/SimplyHTML.jar,
+ .
+Bundle-ActivationPolicy: lazy
+Bundle-Activator: org.flowerplatform.freeplane.FreeplanePlugin
+Export-Package: org.flowerplatform.freeplane.remote
