+++ org.flowerplatform.freeplane/build.properties
@@ -1,0 +1,9 @@
+source.. = src/
+bin.includes = META-INF/,\
+               .,\
+               lib/commons-lang-2.6.jar,\
+               lib/freeplaneeditor_modified.jar,\
+               lib/freeplaneviewer.jar,\
+               lib/gnu-regexp-1.1.4.jar,\
+               lib/jortho.jar,\
+               lib/SimplyHTML.jar
