+++ org.flowerplatform.freeplane/src/org/flowerplatform/freeplane/HeadlessMapViewController.java
@@ -17,7 +17,7 @@
  *  You should have received a copy of the GNU General Public License
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package org.flowerplatform.core.mindmap;
+package org.flowerplatform.freeplane;
 
 import org.freeplane.features.map.IMapSelection;
