+++ org.flowerplatform.freeplane/src/org/flowerplatform/freeplane/controller/FreeplaneAddNodeController.java
@@ -1,8 +1,8 @@
-package org.flowerplatform.core.mindmap;
+package org.flowerplatform.freeplane.controller;
 
-import org.flowerplatform.core.CorePlugin;
 import org.flowerplatform.core.node.controller.AddNodeController;
 import org.flowerplatform.core.node.remote.Node;
+import org.flowerplatform.freeplane.FreeplanePlugin;
 import org.freeplane.features.map.NodeModel;
 
 /**
@@ -12,7 +12,7 @@ public class FreeplaneAddNodeController extends AddNodeController {
 
 	@Override
 	public void addNode(Node node, Node child) {
-		NodeModel parentModel = CorePlugin.getInstance().getFreeplaneUtils().getNodeModel(node.getId());
+		NodeModel parentModel = FreeplanePlugin.getInstance().getFreeplaneUtils().getNodeModel(node.getId());
 		NodeModel newNodeModel = new NodeModel("", parentModel.getMap());
 		newNodeModel.setLeft(false);
