+++ org.flowerplatform.freeplane/src/org/flowerplatform/freeplane/controller/FreeplaneRemoveNodeController.java
@@ -1,8 +1,8 @@
-package org.flowerplatform.core.mindmap;
+package org.flowerplatform.freeplane.controller;
 
-import org.flowerplatform.core.CorePlugin;
 import org.flowerplatform.core.node.controller.RemoveNodeController;
 import org.flowerplatform.core.node.remote.Node;
+import org.flowerplatform.freeplane.FreeplanePlugin;
 import org.freeplane.features.map.NodeModel;
 
 /**
@@ -12,7 +12,7 @@ public class FreeplaneRemoveNodeController extends RemoveNodeController {
 
 	@Override
 	public void removeNode(Node node, Node child) {
-		NodeModel childModel = CorePlugin.getInstance().getFreeplaneUtils().getNodeModel(child.getId());
+		NodeModel childModel = FreeplanePlugin.getInstance().getFreeplaneUtils().getNodeModel(child.getId());
 		childModel.removeFromParent();
 	}
