+++ org.flowerplatform.host.web_app/launch_configs/Flower Platform - Host WebApp (write plugin config to file).launch
@@ -18,9 +18,9 @@
 <stringAttribute key="org.eclipse.jdt.launching.VM_ARGUMENTS" value="-Declipse.ignoreApp=true -Dosgi.noShutdown=true"/>
 <stringAttribute key="pde.version" value="3.3"/>
 <booleanAttribute key="show_selected_only" value="false"/>
-<stringAttribute key="target_bundles" value="javax.xml@default:default,org.apache.felix.gogo.command@default:default,org.apache.felix.gogo.runtime@default:default,org.apache.felix.gogo.shell@default:default,org.eclipse.core.contenttype@default:default,org.eclipse.core.jobs@default:default,org.eclipse.core.runtime@default:true,org.eclipse.equinox.app@default:default,org.eclipse.equinox.common@2:true,org.eclipse.equinox.console@default:default,org.eclipse.equinox.http.registry@default:default,org.eclipse.equinox.http.servlet*1.1.400.v20130418-1354@default:default,org.eclipse.equinox.preferences@default:default,org.eclipse.equinox.registry@default:default,org.eclipse.osgi.services*3.3.100.v20130513-1956@default:default,org.eclipse.osgi@-1:true"/>
+<stringAttribute key="target_bundles" value="javax.xml@default:default,org.eclipse.core.contenttype@default:default,org.eclipse.core.jobs@default:default,org.eclipse.core.runtime@default:true,org.eclipse.equinox.app@default:default,org.eclipse.equinox.common@2:true,org.eclipse.equinox.http.registry@default:default,org.eclipse.equinox.http.servlet@default:default,org.eclipse.equinox.preferences@default:default,org.eclipse.equinox.registry@default:default,org.eclipse.osgi.services@default:default,org.eclipse.osgi@-1:true"/>
 <booleanAttribute key="tracing" value="false"/>
 <booleanAttribute key="useCustomFeatures" value="false"/>
 <booleanAttribute key="useDefaultConfigArea" value="true"/>
-<stringAttribute key="workspace_bundles" value="org.flowerplatform.core@default:default,org.flowerplatform.flex_client.core@default:default,org.flowerplatform.flex_client.host_app.config.web@default:default,org.flowerplatform.flex_client.host_app@default:default,org.flowerplatform.flex_client.properties@default:default,org.flowerplatform.flex_client.server@default:default,org.flowerplatform.flex_client.web@default:default,org.flowerplatform.host.servletbridge.extensionbundle@default:false,org.flowerplatform.host.servletbridge.registry@default:default,org.flowerplatform.util@default:default"/>
+<stringAttribute key="workspace_bundles" value="org.flowerplatform.core@default:default,org.flowerplatform.flex_client.core@default:default,org.flowerplatform.flex_client.host_app.config.web@default:default,org.flowerplatform.flex_client.host_app@default:default,org.flowerplatform.flex_client.properties@default:default,org.flowerplatform.flex_client.server@default:default,org.flowerplatform.flex_client.web@default:default,org.flowerplatform.freeplane@default:default,org.flowerplatform.host.servletbridge.extensionbundle@default:false,org.flowerplatform.host.servletbridge.registry@default:default,org.flowerplatform.util@default:default"/>
 </launchConfiguration>
